pyproject.toml
[project]
name = "secure-with-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "diskcache>=5.6.3",
    "google-genai>=1.50.1",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.38.0",
    "streamlit>=1.51.0",
]
